LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/text/2d.text.measure.fontBoundingBox-expected.txt

2d.text.measure.fontBoundingBox

Testing fontBoundingBox measurements


FAIL Testing fontBoundingBox measurements assert_equals: ctx.measureText('A').fontBoundingBoxAscent === 30 (got 68[number], expected 30[number]) expected 30 but got 68
